Compiled App, Actual calculations.
I made it so it functioned as an actual calculator, and you can calculate things on it. I would go more into detail, but for what? It's a calculator, it does what's on the tin. I added the UI so it's viewable and also compiled it into an EXE so that you can run it.